==> pipeline_11_horas/pipeline/notebooks_respaldo/1_concat_pronostico_contaminacion.ipynb <==
import pandas as pd

path_inicial_1 = "./datos/ingesta_pronostico_meteo.csv"
path_inicial_2 = "./datos/ingesta_contaminacion.csv"
path_final = "./datos/ingesta_contaminacion.csv"

df1 =  pd.read_csv(path_inicial_1)
df2 = pd.read_csv(path_inicial_2) 

df = pd.concat( [ df1, df2 ],sort=True )

df = df.sort_values(["id_station","fecha"],ascending=False)

df = df.rename({"PM2":"PM2.5"},axis=1)

df.to_csv(path_final,index=False)
