==> pipeline_11_horas/pipeline/notebooks_respaldo/2-ingesta_datos_limpieza.ipynb <==
## Juntar y limpiar archivos

En este notebook vamos a limpiar y a generar groupbys en los archivos históricos para poder alimentar a los modelos de ML adecuadamente.

- __Datos recibidos:__ obtenidos de [el portal de calidad del aire de la CDMX](http://www.aire.cdmx.gob.mx/default.php?opc='aKBhnmI='&opcion=Zg==)
- __Responsable:__ Daniel Bustillos
- __Contacto:__  juandaniel.bucam@gmail.com

__Notas del proyecto__
Necesitamos generar un csv con los niveles máximos, mínimo y promedio por día, se perderá la información de la estación, también generaremos el atributo PM10mean y PM25mean que representarán los valores de las últimas 24 horas de los contaminantes.

calidad de caire
pronostico contaminación

## Pasos
[X] Leer los 10 archivos de contaminación <br>
[X] concatenar los archivos<br>
[X] Hacer el mes y el día columnas<br><br>
[X] Hacer el rollinng para PM10 y PM25<br><br>
[X] identidicar y eliminar Outliers
[X] Formatear Fechas
juntar<br>
[X] juntar archivos de append y rolling <br><br>


import numpy as np


import pandas as pd

import datetime

import numpy as np
import matplotlib.pyplot as plt

import seaborn as sns; sns.set()
import matplotlib.pyplot as plt

pd.options.mode.chained_assignment = None 

path_inicial = "./datos/ingesta_contaminacion.csv"
path_final ="./datos/ingesta_hora_limpios.csv"

## Leamos el DF

Leemos los archivos por hora:

Declaramos todos los archivos enuna sola lista

df_append = pd.read_csv(path_inicial)

df_append

# df_append = #pd.read_csv("/home/paw/DanielBustillos/contaminación/notebooks_streaming/csv's/dataset_ingesta.csv")

#df_append.head()

df_append['fecha'] = pd.to_datetime(df_append['fecha'], format='%Y-%m-%d %H:%M:%S')

Ordenemos el DF por fecha y estación:

df_append = df_append.sort_values(['fecha','id_station'], ascending=[-1, 1])

Vamos a quedarnos con las celdas donde Pm10 y Pm2.5 son no nulas.

df_append

#df_append = df_append.dropna() ##porque????
#df_append = df_append.dropna(subset=["PM10"]).reset_index()
#df_append["PM2.5"] = df_append["PM2.5"].astype(float)
#df_append["NO"] = df_append["NO"].astype(float)

### Creamos un Df con los datos promedio por hora oara rellenar los nan del DF original

data_24 = df_append.set_index('fecha')

ahora vamos a sacar el promedio de las ultimas 24 horas por estación para el PM10 y el PM25:

data_24 = data_24.groupby('id_station')[["PM10","PM2.5"]].rolling(24).mean().dropna().reset_index()

data_24 = data_24.rename(columns={'PM10': 'PM10mean',"PM2.5":'PM25mean'})

## Unir los archivos:

Unimos los dos DF:

data_day = pd.merge(df_append, data_24, on=["fecha","id_station"],how="left")

data_day[["fecha","PM10","PM10mean"]][data_day.id_station=="BJU"].reset_index(drop=True).tail(60)

Veamos PM10mean y PM25mean

Vamos a eliminar las estaciones PED y PLA, aparentemente tienen muchos outliers:

#data_day = data_day[data_day.id_station!="PED"]
#data_day = data_day[data_day.id_station!="PLA"]

data_day.to_csv(path_final, index=False)
