==> pipeline_11_horas/pipeline/notebooks_respaldo/3_ingesta_datos_shift_O3.ipynb <==
# Datos entreneamiento Shift

En este notebook vamos a aplicar el shift para las obtener las columnas mas correlacionadas.

- __Datos recibidos:__ obtenidos de notebook __1-datos_entrenamiento_shift.ipynb__
- __Responsable:__ Daniel Bustillos
- __Contacto:__  juandaniel.bucam@gmail.com

__Notas del proyecto__
Necesitamos generar un csv con los niveles máximos, mínimo y promedio por día, se perderá la información de la estación, también generaremos el atributo PM10mean y PM25mean que representarán los valores de las últimas 24 horas de los contaminantes.

calidad de caire
pronostico contaminación

## Pasos
[X] Aplicar Shift con Correlaciones <br>
[X] Sacar minimo promedio y maximo <br>
[X] Obtener columnas más correlacionadas<br><br>

import pandas as pd
pd.options.mode.chained_assignment = None 

path_inicial = "/home/paw/DanielBustillos/contaminación/pipeline-norberto/pipeline_pau/datos/ingesta_hora_limpios.csv"
path_final = "/home/paw/DanielBustillos/contaminación/pipeline-norberto/pipeline_pau/datos/ingesta_modelos_O3.csv"

data_hour_merge_24 = pd.read_csv(path_inicial)

data_hour_merge_24.columns

Quedémonos sin WSR:

data_hour_merge_24 = data_hour_merge_24[['CO', 'NO', 'NO2', 'NOX', 'O3', 'PM10', "PM2.5",'RH', 'SO2', 'TMP',
       'WSP', 'dia', 'fecha', 'hora', 'id_station', 'mes',"PM10mean","PM25mean"]].reset_index(drop=True)

Hagamos una lista con todas las estaciones:

### Arreglemos las fechas:

data_hour_merge_24['dia'] = data_hour_merge_24['dia'].astype(str).astype(float)
data_hour_merge_24['mes'] = data_hour_merge_24['mes'].astype(str).astype(float)

Creemos un diccionario con __data_hour_merge_24__ dividido por estación:

estaciones = data_hour_merge_24.id_station.unique().tolist()

data_est = {}
for elem in estaciones:
    data_est[elem] = data_hour_merge_24[data_hour_merge_24.id_station == elem]

# Leamos el archivo de correlaciones del PM10 mean

Este archivo se generó en el notebook: __/home/paw/DanielBustillos/contaminación/correlaciones_pau/correlaciones_función_paulina.ipynb__

target = "O3_y"

O3_corr = pd.read_csv("/DATA/paw/jupyterhub_notebook/daniel.bustillos/DanielBustillos/contaminación/pipeline-norberto/correlacion/correlaciones_"+target+".csv")
O3_corr = O3_corr.iloc[:,1:]
O3_corr.head(5)

Vamos a quedarnos con las variables con corr>0.3:

O3_corr_filtro = O3_corr[(O3_corr.valor>0.4) | (O3_corr.valor<-0.4) ].reset_index(drop=True)

O3_corr_filtro.shape

Vamos a aplicar el shift para cada elemento de la tabla __O3_corr__:

def shit_corr(df):
    for i in range(len(O3_corr_filtro)):
        name_column = str( O3_corr_filtro.loc[i,"contaminante"] + "_" + str( O3_corr_filtro.loc[i,"horas"] ) ) 
        df[name_column] = df[O3_corr_filtro.loc[i,"contaminante"]].shift( - int(float(str(O3_corr_filtro.loc[i,"horas"]))) ) 
    return df.dropna()

Aplicamos esta función para cada DF de cada estación:

data_shift = {}
for elem in data_est:
    data_shift[elem] = shit_corr(data_est[elem])

Creamos un Df vacío:

df_append = pd.DataFrame(columns = data_est["MER"].columns.tolist())

for key in data_est:
    df_append = df_append.append(data_est[key], ignore_index=True,sort=True)

## Groupbys

cols = df_append.columns.tolist()

cols.remove("id_station")

Finalmente, aplicamos los groupbys:

data_hour_merge_24_mean = df_append.groupby(['fecha',"hora","dia","mes"]).mean()
data_hour_merge_24_mean.reset_index(inplace=True)

data_hour_merge_24_max = df_append.groupby(['fecha',"hora","dia","mes"]).max()
data_hour_merge_24_max.reset_index(inplace=True)

data_hour_merge_24_min = df_append.groupby(['fecha',"hora","dia","mes"]).min()
data_hour_merge_24_min.reset_index(inplace=True)

### Arreglemos los nombres:

max_columns = data_hour_merge_24_max.columns.tolist()
min_columns = data_hour_merge_24_min.columns.tolist()
mean_columns = data_hour_merge_24_mean.columns.tolist()

for i in range(len(max_columns)):
    if max_columns[i] not in ['fecha','hora','dia','mes','id_station']:
        max_columns[i] = max_columns[i]+"_max"
        min_columns[i] = min_columns[i]+"_min"    
        mean_columns[i] = mean_columns[i]+"_mean"

data_hour_merge_24_mean.columns = mean_columns
data_hour_merge_24_min.columns = min_columns
data_hour_merge_24_max.columns = max_columns

##### Unamos los df's de manera que tengamos en una solo los datos promedio, máximo y minimo por día:

data_hour_merge = pd.merge(data_hour_merge_24_mean, data_hour_merge_24_max, on=['fecha',"hora","dia","mes"])
data_hour_merge = pd.merge(data_hour_merge, data_hour_merge_24_min, on=['fecha',"hora","dia","mes"])

data_hour_merge = data_hour_merge.sort_values(['fecha',"hora"], ascending=[0,1])

## Leamos el archivo:

## Correlacionados con el target.

Guardemos estas variables, no pueden cambiarse una vez que los modelos estan entrenados:

Filtremos el DF con las variables mas correlacionadas:

O3_corr = ['O3_25.0_mean',  'O3_22.0_mean',  'TMP_48.0_mean',  'RH_mean',  'O3_max',  'TMP_1.0_min',  
           'O3_0.0_max',  'O3_44.0_min',  'O3_45.0_max',  'O3_21.0_min',  'hora_10.0_mean',  
           'TMP_23.0_mean',  'RH_min',  'hora_13.0_mean',  'O3_0.0_mean',  'O3_44.0_max', 
           'RH_24.0_mean',  'hora_13.0_max',  'O3_20.0_mean',  'O3_0.0_min',  'RH_max', 
           'O3_24.0_mean',  'TMP_21.0_mean',  'O3_21.0_max',  'O3_mean',  'O3_22.0_min', 
           'RH_45.0_mean',  'TMP_max',  'O3_23.0_max',  'O3_2.0_mean',  'TMP_47.0_max', 
           'O3_47.0_max',  'O3_2.0_max',  'TMP_24.0_mean',  'TMP_23.0_max',  'RH_22.0_min', 
           'O3_3.0_max',  'O3_21.0_mean',  'O3_47.0_mean',  'hora_11.0_mean',  'TMP_0.0_max', 
           'RH_2.0_min',  'RH_21.0_mean',  'RH_21.0_min',  'hora_11.0_max',  'hora_12.0_min', 
           'O3_48.0_mean',  'RH_22.0_max',  'RH_25.0_mean',  'O3_1.0_min',  'O3_46.0_max', 
           'RH_0.0_mean',  'O3_25.0_max',  'O3_2.0_min',  'RH_45.0_min',  'RH_46.0_mean', 
           'TMP_25.0_mean',  'TMP_0.0_min',  'O3_1.0_max',  'RH_23.0_mean',  'hora_36.0_max', 
           'RH_22.0_mean',  'O3_24.0_max',  'TMP_0.0_mean',  'RH_1.0_max',  'TMP_22.0_mean',  
           'WSP_mean',  'TMP_22.0_max',  'TMP_25.0_max',  'TMP_1.0_mean',  'O3_46.0_min', 
           'hora_12.0_max',  'hora_36.0_mean',  'O3_45.0_min',  'O3_46.0_mean',  'O3_1.0_mean', 
           'hora_34.0_mean',  'TMP_2.0_max',  'TMP_mean',  'TMP_1.0_max',  'TMP_3.0_max',  
           'TMP_47.0_mean',  'RH_0.0_max',  'RH_46.0_min',  'TMP_44.0_mean',  'O3_min', 
           'TMP_45.0_mean',  'O3_22.0_max',  'TMP_24.0_max',  'O3_23.0_min',  'RH_0.0_min', 
           'RH_1.0_min',  'TMP_46.0_mean',  'O3_44.0_mean',  'hora_11.0_min',  'RH_23.0_min', 
           'RH_2.0_mean',  'TMP_21.0_max',  'RH_1.0_mean',  'TMP_46.0_max',  'RH_24.0_min',  
           'O3_45.0_mean',  'TMP_2.0_mean',  'RH_47.0_mean',  'O3_24.0_min',  'hora_35.0_mean',  
           'hora_12.0_mean',  'O3_20.0_min',  'O3_20.0_max',  'O3_23.0_mean',  'TMP_45.0_max', 
           'TMP_min',  'RH_44.0_mean',  'RH_mean_frcst_1',  'RH_mean_frcst_2',  'RH_mean_frcst_3', 
           'RH_mean_frcst_4',  'RH_mean_frcst_5',  'RH_mean_frcst_6',  'RH_mean_frcst_7',  
           'RH_mean_frcst_8',  'RH_mean_frcst_9',  'RH_mean_frcst_10',  'RH_mean_frcst_11', 
           'RH_mean_frcst_12',  'RH_mean_frcst_13',  'RH_mean_frcst_14',  'RH_mean_frcst_15',  
           'RH_mean_frcst_16',  'RH_mean_frcst_17',  'RH_mean_frcst_18',  'RH_mean_frcst_19', 
           'RH_mean_frcst_20',  'RH_mean_frcst_21',  'RH_mean_frcst_22',  'RH_mean_frcst_23',  
           'RH_mean_frcst_24',  'WSP_mean_frcst_1',  'WSP_mean_frcst_2',  'WSP_mean_frcst_3', 
           'WSP_mean_frcst_4',  'WSP_mean_frcst_5',  'WSP_mean_frcst_6',  'WSP_mean_frcst_7',  
           'WSP_mean_frcst_8',  'WSP_mean_frcst_9',  'WSP_mean_frcst_10',  'WSP_mean_frcst_11',
           'WSP_mean_frcst_12',  'WSP_mean_frcst_13',  'WSP_mean_frcst_14',  'WSP_mean_frcst_15', 
           'WSP_mean_frcst_16',  'WSP_mean_frcst_17',  'WSP_mean_frcst_18',  'WSP_mean_frcst_19', 
           'WSP_mean_frcst_20',  'WSP_mean_frcst_21',  'WSP_mean_frcst_22',  'WSP_mean_frcst_23', 
           'WSP_mean_frcst_24',  'TMP_mean_frcst_1',  'TMP_mean_frcst_2',  'TMP_mean_frcst_3',
           'TMP_mean_frcst_4',  'TMP_mean_frcst_5',  'TMP_mean_frcst_6',  'TMP_mean_frcst_7',  
           'TMP_mean_frcst_8',  'TMP_mean_frcst_9',  'TMP_mean_frcst_10',  'TMP_mean_frcst_11', 
           'TMP_mean_frcst_12',  'TMP_mean_frcst_13',  'TMP_mean_frcst_14',  'TMP_mean_frcst_15',
           'TMP_mean_frcst_16',  'TMP_mean_frcst_17',  'TMP_mean_frcst_18',  'TMP_mean_frcst_19',  
           'TMP_mean_frcst_20',  'TMP_mean_frcst_21',  'TMP_mean_frcst_22',  'TMP_mean_frcst_23', 
           'TMP_mean_frcst_24']

O3_corr = ["fecha",'hora', 'NO_x', 'NOX_x', 'O3_x', 'O3_0_x', 'O3_1_x', 'O3_10_x', 'O3_11_x', 'O3_12_x', 'O3_13_x', 
           'O3_18_x', 'O3_2_x', 'O3_3_x', 'O3_8_x', 'O3_9_x', 'PM10_x', 'PM10mean_x', 'PM25mean_x', 'RH_x',
           'TMP_x', 'TMP_0_x', 'TMP_1_x', 'TMP_10_x', 'TMP_11_x', 'TMP_12_x', 'TMP_13_x', 'TMP_14_x', 
           'TMP_2_x', 'TMP_3_x', 'TMP_4_x', 'TMP_9_x', 'WSP_x', 'NO_y', 'NOX_y', 'O3_y', 'O3_0_y', 'O3_1_y', 
           'O3_18_y', 'O3_2_y', 'O3_3_y', 'O3_4_y', 'PM10mean_y', 'PM25mean_y', 'RH_y', 'TMP_y', 'TMP_0_y',
           'TMP_1_y', 'TMP_10_y', 'TMP_11_y', 'TMP_12_y', 'TMP_13_y', 'TMP_14_y', 'TMP_2_y', 'TMP_3_y', 
           'TMP_4_y', 'TMP_9_y', 'WSP_y', 'NO', 'O3', 'O3_0', 'O3_1', 'O3_10', 'O3_11', 'O3_12', 'O3_13', 
           'O3_18', 'O3_2', 'O3_3', 'O3_7', 'O3_8', 'O3_9', 'PM10', 'PM10mean', 'PM2.5_y', 'PM25mean', 'RH',
           'TMP', 'TMP_0', 'TMP_1', 'TMP_10', 'TMP_11', 'TMP_12', 'TMP_13', 'TMP_14', 'TMP_2', 'TMP_3', 'TMP_9', 'WSP'] para modelos pasados

Añadamos los datos de pronostico de las siguientes variables:

lista_frcst = ["RH_mean","WSP_mean","TMP_mean"]

for item in lista_frcst:
    for i in range(1, 25):
        col_name = str( item + "_frcst_" + str(i))
        data_hour_merge[col_name] = data_hour_merge[item].shift(i)
#data_hour_merge.dropna(inplace=True)

import datetime
#inicio= 46 - datetime.datetime.now().hour + 1

#data_hour_merge = data_hour_merge.reset_index(drop=True).iloc[inicio:400,:].fillna(30)

data_hour_merge = data_hour_merge.tail(-24)
data_hour_merge = data_hour_merge.fillna(data_hour_merge.mean())

data_hour_merge = data_hour_merge[O3_corr]

Guardemos:

data_hour_merge.to_csv(path_final,sep=',', encoding='utf-8',index=False)

data_hour_merge.columns.shape
